% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\name{fb_auth}
\alias{fb_auth}
\title{Authorize Facebook API with your token}
\usage{
fb_auth(
  client_id = Sys.getenv("FACEBOOK_CLIENT_ID"),
  client_secret = Sys.getenv("FACEBOOK_CLIENT_SECRET"),
  token = NULL,
  new_user = FALSE,
  silent = FALSE
)
}
\description{
Enter your token that can be found on https://developers.facebook.com/.
}
\examples{
fb_auth(client_id = Sys.getenv("FACEBOOK_CLIENT_ID"), client_secret = Sys.getenv("FACEBOOK_CLIENT_SECRET"), token = NULL, new_user = FALSE, silent = FALSE)
}
