% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_query.R
\name{fb_query}
\alias{fb_query}
\title{Get custom API request string}
\usage{
fb_query(query)
}
\arguments{
\item{query}{Query parameters for the Graph API}
}
\description{
Query Facebook API with json encoded string
}
\examples{
fb_query("{me{name, id}}")
}
